Add Refund#perform_response
There are instances when we need to be able to access the payment
gateway response received during the refund perform process, notably
in core/app/models/spree/payment/cancellation.rb:31 where it assumes
that #try_void returns the gateway response:

```
  if response = payment.payment_method.try_void(payment)
    payment.send(:handle_void_response, response)
  else
    #...
  end
```

the next line calls core/app/models/spree/payment/processing.rb:191
which makes super clear that we're expecting an object that responds
to #success? and #authorization, ie. the gateway response.

```
  def handle_void_response(response)
    record_response(response)

    if response.success?
      self.response_code = response.authorization
      void
    else
      gateway_error(response)
    end
  end
```