Capture and void only payments with positive amount
There's no point in capturing or voiding a payment with zero amount.

When capturing zero amount, ActiveMerchant will respond with the
error `This value must be greater than or equal to 1.`, while if
payment amount is zero, the error from ActiveMerchant responds
with the error `Amount must be at least $0.50 usd`.