Add `Event::SubscriberRegistry` class
The class `Spree::Event::SubscriberRegistry` is introduced in
order to centralize and handle in a dedicated environment
event mappings for subscriber modules. This was previously
responsibility of each `Spree::Event::Subscriber` module, but
due to some issues with Zeitwerk and the way these modules
were activated in rails initializers, that approach is not
working properly anymore with Rails 6.

This class also manages the activation/deactivation of all
event subscribers that happens at boot time.

The activation/deactivation code is threadsafe, just in the
remote case that somebody tries to use it in a multithreaded
context. This does not happen in the standard use-case of
this class, as it is being used during the app initialization
time.