Test refunds controller stubbing process! instead of perform!
The first line of perform! avoid the method to be executed under
certain circumstances (which are actually happening in the spec
context). Due to this, we are not testing the current behavior
since perform! should be skipped and in the spec it's raising
an exception.

To fix this, it's better to stub process!, which is where the
simulated exception can happen. We now have a failing spec but
it's reflecting what happens in the real world.